+++ src/components/site-form.js
@@ -108,7 +108,19 @@ const createSiteFormTemplate = (luggage, comfort, meal, seat, train, description
         </div>
 
         <button class="event__save-btn  btn  btn--blue" type="submit">Save</button>
-        <button class="event__reset-btn" type="reset">Cancel</button>
+        <button class="event__reset-btn" type="reset">Delete</button>
+        
+        <input id="event-favorite-1" class="event__favorite-checkbox  visually-hidden" type="checkbox" name="event-favorite" checked>
+        <label class="event__favorite-btn" for="event-favorite-1">
+          <span class="visually-hidden">Add to favorite</span>
+          <svg class="event__favorite-icon" width="28" height="28" viewBox="0 0 28 28">
+            <path d="M14 21l-8.22899 4.3262 1.57159-9.1631L.685209 9.67376 9.8855 8.33688 14 0l4.1145 8.33688 9.2003 1.33688-6.6574 6.48934 1.5716 9.1631L14 21z"/>
+          </svg>
+        </label>
+        
+         <button class="event__rollup-btn" type="button">
+            <span class="visually-hidden">Open event</span>
+         </button>
       </header>
       <section class="event__details">
         <section class="event__section  event__section--offers">
@@ -199,6 +211,10 @@ export class Form extends AbstractComponent {
   }
 
   setClickHandler(handler) {
-    this.getElement().addEventListener(`submit`, handler);
+    this.getElement().querySelector(`.event__save-btn`).addEventListener(`click`, handler);
+  }
+
+  setClickStarHandler(handler) {
+    this.getElement().querySelector(`.event__favorite-checkbox`).addEventListener(`click`, handler);
   }
 }
